server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "joblib>=1.4.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-levenshtein>=0.25.0",
    "requests>=2.32.5",
    "scikit-learn>=1.6.1,<1.7",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
]
